relocating flamegraph view to app
