+++ x-pack/plugins/profiling/public/components/app/flame_graphs_view/flame_graph_legend.tsx
@@ -7,8 +7,8 @@
 import { EuiFlexGroup, EuiFlexItem, EuiText } from '@elastic/eui';
 import React from 'react';
 import { i18n } from '@kbn/i18n';
-import { asPercentage } from '../../utils/formatters/as_percentage';
-import { Legend, LegendItem } from '../legend';
+import { asPercentage } from '../../../utils/formatters/as_percentage';
+import { Legend, LegendItem } from '../../legend';
 
 export function FlameGraphLegend({
   legendItems,
