+++ x-pack/plugins/profiling/public/components/app/flame_graphs_view/flamegraph_information_window.tsx
