+++ x-pack/plugins/profiling/public/components/app/flame_graphs_view/index.tsx
@@ -17,19 +17,22 @@ import {
 import { i18n } from '@kbn/i18n';
 import { get } from 'lodash';
 import React, { useState } from 'react';
-import { FlameGraphComparisonMode, FlameGraphNormalizationMode } from '../../../common/flamegraph';
-import { useProfilingParams } from '../../hooks/use_profiling_params';
-import { useProfilingRouter } from '../../hooks/use_profiling_router';
-import { useProfilingRoutePath } from '../../hooks/use_profiling_route_path';
-import { useTimeRange } from '../../hooks/use_time_range';
-import { useTimeRangeAsync } from '../../hooks/use_time_range_async';
-import { AsyncComponent } from '../async_component';
-import { useProfilingDependencies } from '../contexts/profiling_dependencies/use_profiling_dependencies';
-import { FlameGraph } from '../flamegraph';
-import { PrimaryAndComparisonSearchBar } from '../primary_and_comparison_search_bar';
-import { PrimaryProfilingSearchBar } from '../profiling_app_page_template/primary_profiling_search_bar';
-import { ProfilingAppPageTemplate } from '../profiling_app_page_template';
-import { RedirectTo } from '../redirect_to';
+import {
+  FlameGraphComparisonMode,
+  FlameGraphNormalizationMode,
+} from '../../../../common/flamegraph';
+import { useProfilingParams } from '../../../hooks/use_profiling_params';
+import { useProfilingRouter } from '../../../hooks/use_profiling_router';
+import { useProfilingRoutePath } from '../../../hooks/use_profiling_route_path';
+import { useTimeRange } from '../../../hooks/use_time_range';
+import { useTimeRangeAsync } from '../../../hooks/use_time_range_async';
+import { AsyncComponent } from '../../async_component';
+import { useProfilingDependencies } from '../../contexts/profiling_dependencies/use_profiling_dependencies';
+import { FlameGraph } from '../../flamegraph';
+import { PrimaryAndComparisonSearchBar } from '../../primary_and_comparison_search_bar';
+import { PrimaryProfilingSearchBar } from '../../profiling_app_page_template/primary_profiling_search_bar';
+import { ProfilingAppPageTemplate } from '../../profiling_app_page_template';
+import { RedirectTo } from '../../redirect_to';
 import { FlameGraphNormalizationOptions, NormalizationMenu } from './normalization_menu';
 
 export function FlameGraphInformationWindowSwitch({
