+++ x-pack/plugins/profiling/public/routing/index.tsx
@@ -12,7 +12,7 @@ import React from 'react';
 import { TopNFunctionSortField, topNFunctionSortFieldRt } from '../../common/functions';
 import { StackTracesDisplayOption, TopNType } from '../../common/stack_traces';
 import { FlameGraphComparisonMode, FlameGraphNormalizationMode } from '../../common/flamegraph';
-import { FlameGraphsView } from '../components/flame_graphs_view';
+import { FlameGraphsView } from '../components/app/flame_graphs_view';
 import { FunctionsView } from '../components/functions_view';
 import { RedirectTo } from '../components/redirect_to';
 import { RouteBreadcrumb } from '../components/route_breadcrumb';
